Work around Composer multi-version loading hack using a "V2" namespace
Composer would try to load two versions of the same plugin, by reading
the main plugin file and rewriting the class name. But this doesn't work
if the plugin has multiple classes. It tries to run the new plugin class
with old helper classes.

So:

* Move all classes to Wikimedia\Composer\Merge\V2. This includes classes
  that were previously in the parent namespace.
* Clean up unused "use" statements.

Note that the namespace version number will have to be incremented every
time there is an internal interface change.

Closes #205